Add setup script + new cabal build style
The setup script assumes that void-extra-packages and void-packages are
checked out side-by-side in the same directory, and will copy things
from one into the other. (We can't use symlinks, because xbps-pkg thinks
that symlinks are a sign of a subpackage!)

This also adds the haskell-cabal build style, which abstracts out the
process of setting up a sandbox and installing things with Cabal.